requirements.txt
# Pakai PyTorch yang sudah ada di Kaggle (>=2.0, biasanya 2.5.0)
torch>=2.0
torchvision>=0.15
torchaudio>=2.0

# Lib tambahan
numpy>=1.16.5
scipy>=1.3.0
opencv-python
gdown
pillow
gradio
timm==0.4.12   # versi ini yang dipakai repo kamu
wandb
matplotlib
